fix(api): refactor updatepatientdiscover to be shared
Ref. metriport/metriport-internal#1350
