+++ packages/api/src/external/commonwell/patient-external-data.ts
@@ -97,59 +97,3 @@ export const setCommonwellId = async ({
     return updatedPatient.update({ data: updatedData }, { transaction });
   });
 };
-
-/**
- * Sets the CommonWell (CW) integration status on the patient.
- *
- * @param patientId The patient ID @ Metriport.
- * @param cxId The customer ID @ Metriport.
- * @param status The status of integrating/synchronizing the patient @ CommonWell.
- * @returns
- */
-export const setPatientDiscoveryStatus = async ({
-  patientId,
-  cxId,
-  status,
-}: {
-  patientId: string;
-  cxId: string;
-  status: LinkStatus;
-}): Promise<Patient> => {
-  const patientFilter = {
-    id: patientId,
-    cxId,
-  };
-
-  return await executeOnDBTx(PatientModel.prototype, async transaction => {
-    const existingPatient = await getPatientOrFail({
-      ...patientFilter,
-      lock: true,
-      transaction,
-    });
-
-    const externalData = existingPatient.data.externalData ?? {};
-
-    const updatePatientDiscoveryStatus = {
-      ...externalData,
-      COMMONWELL: {
-        ...externalData.COMMONWELL,
-        status,
-      },
-    };
-
-    const updatedPatient = {
-      ...existingPatient.dataValues,
-      data: {
-        ...existingPatient.data,
-        externalData: updatePatientDiscoveryStatus,
-      },
-    };
-
-    await PatientModel.update(updatedPatient, {
-      where: patientFilter,
-      transaction,
-    });
-
-    return updatedPatient;
-  });
-};
