+++ packages/api/src/external/commonwell/patient-shared.ts
@@ -33,7 +33,7 @@ export class PatientDataCommonwell extends PatientExternalDataEntry {
   constructor(
     public patientId: string,
     public personId?: string | undefined,
-    public status?: LinkStatus | undefined,
+    public discoverStatus?: LinkStatus | undefined,
     public cqLinkStatus?: CQLinkStatus,
     public scheduledDocQueryRequestId?: string | undefined
   ) {
