+++ packages/api/src/external/hie/update-patient-discovery-status.ts
@@ -1,14 +1,17 @@
 import { Patient } from "@metriport/core/domain/patient";
-import { getPatientOrFail } from "../../../command/medical/patient/get-patient";
-import { PatientModel } from "../../../models/medical/patient";
-import { executeOnDBTx } from "../../../models/transaction-wrapper";
+import { MedicalDataSource } from "@metriport/core/external/index";
+import { getPatientOrFail } from "../../command/medical/patient/get-patient";
+import { PatientModel } from "../../models/medical/patient";
+import { executeOnDBTx } from "../../models/transaction-wrapper";
 
 export async function updatePatientDiscoveryStatus({
   patient,
   status,
+  source,
 }: {
   patient: Pick<Patient, "id" | "cxId">;
   status: "processing" | "completed" | "failed";
+  source: MedicalDataSource;
 }): Promise<Patient> {
   const patientFilter = {
     id: patient.id,
@@ -26,8 +29,8 @@ export async function updatePatientDiscoveryStatus({
 
     const updatePatientDiscoveryStatus = {
       ...externalData,
-      CAREQUALITY: {
-        ...externalData.CAREQUALITY,
+      [source]: {
+        ...externalData[source],
         discoveryStatus: status,
       },
     };
